+++ test/sampleTest.test.js
@@ -2,7 +2,7 @@
  * NOTE: To run specifically, ENV=<environment> npm run test -- --<reporter> ./test/<testFileName>
 **/
 
-homePage = require('../pages/home.page')
+let homePage = require('../pages/home.page')
 
 describe('Sample Test - Go To Sign Up Page', function() {
     it('should take you to the sign up page', () => {
